# devices-service/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Devices Service API",
    "version": "1.0.0"
  },
  "paths": {
    "/devices/{device_id}": {
      "get": {
        "summary": "Get Device Info",
        "description": "Returns detailed information about a specific device by its ID.",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
    "/devices/{device_id}/status": {
      "put": {
        "summary": "Update Device Status",
        "description": "Allows changing the status of a device.",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
    "/devices/{device_id}/commands": {
      "post": {
        "summary": "Send Command to Device",
        "description": "Sends a command to the device.",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "command": {
                    "type": "string"
                  },
                  "value": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command sent"
          },
          "400": {
            "description": "Invalid command"
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    }
  }
}